Move tests into tests/ directory and add vitest config
- Move all 11 test files from src/ to tests/
- Add vitest.config.ts with explicit test include pattern
- Update imports in test files to reference ../src/ modules
- Update tsconfig.json to include tests/ and vitest.config.ts
- Simplify package.json files field (test exclusion no longer needed)

Fixes #10
